+++ amethyst_core/src/lib.rs
@@ -21,6 +21,7 @@ extern crate thread_profiler;
 #[cfg(all(target_os = "emscripten", not(no_threading)))]
 compile_error!("the cfg flag \"no_threading\" is required when building for emscripten");
 
+pub use self::named::{Named, WithNamed};
 pub use bundle::{Error, ErrorKind, Result, SystemBundle};
 pub use orientation::Orientation;
 pub use timing::*;
@@ -30,6 +31,7 @@ use std::sync::Arc;
 
 pub mod bundle;
 pub mod frame_limiter;
+mod named;
 mod orientation;
 pub mod timing;
 pub mod transform;