+++ examples/pong_tutorial_01/main.rs
@@ -2,7 +2,9 @@ extern crate amethyst;
 
 use amethyst::input::{is_close_requested, is_key_down};
 use amethyst::prelude::*;
-use amethyst::renderer::{DrawFlat, Event, PosTex, VirtualKeyCode};
+use amethyst::renderer::{
+    DisplayConfig, DrawSprite, Event, Pipeline, RenderBundle, Stage, VirtualKeyCode,
+};
 
 pub struct Pong;
 
@@ -28,13 +30,20 @@ fn main() -> amethyst::Result<()> {
         "{}/examples/pong_tutorial_01/resources/display_config.ron",
         env!("CARGO_MANIFEST_DIR")
     );
+    let config = DisplayConfig::load(&path);
+
+    let pipe = Pipeline::build().with_stage(
+        Stage::with_backbuffer()
+            .clear_target([0.0, 0.0, 0.0, 1.0], 1.0)
+            .with_pass(DrawSprite::new()),
+    );
+    let game_data = GameDataBuilder::default()
+        .with_bundle(RenderBundle::new(pipe, Some(config)).with_sprite_sheet_processor())?;
 
     // This line is not mentioned in the pong tutorial as it is specific to the context
     // of the git repository. It only is a different location to load the assets from.
     let assets_dir = format!("{}/examples/assets/", env!("CARGO_MANIFEST_DIR"));
 
-    let game_data =
-        GameDataBuilder::default().with_basic_renderer(path, DrawFlat::<PosTex>::new(), false)?;
     let mut game = Application::new(assets_dir, Pong, game_data)?;
     game.run();
     Ok(())